refactor(#295): 관리자 관련 API URI 개선
