+++ src/main/java/com/sideproject/cafe_cok/menu/presentation/MenuController.java
@@ -5,11 +5,9 @@ import com.sideproject.cafe_cok.menu.application.MenuService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/menu")
@@ -26,4 +24,12 @@ public class MenuController {
         MenuListResponse response = menuService.findByCafeId(cafeId);
         return ResponseEntity.ok(response);
     }
+
+    @DeleteMapping("/{menuId}")
+    public ResponseEntity<String> delete(@PathVariable Long menuId) {
+        boolean success = menuService.delete(menuId);
+        if(success) return ResponseEntity.ok("메뉴가 성공적으로 삭제되었습니다.");
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("메뉴 삭제에 실패했습니다.");
+    }
+
 }
